+++ lang/cs/navig.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+
+    'login' => 'Přihlášení',
+    'register' => 'Nový účet',
+    'register.email' => 'Nový účet s emailem',
+    'users.index' => 'Uživatelé',
+    'users.show' => 'Detail uživatele',
+    'pin-login' => 'Přihlášení PINem',
+    'pin-login.verify' => 'Ověření PINem',
+
+];